convert page to server component that renders exisiting page as client component
